pyproject.toml
[project]
name = "pageflip"
version = "0.1.0"
description = "Capture slides from a screen region whenever they change"
readme = "README.md"
requires-python = ">=3.11"
license = { text = "Apache-2.0" }
authors = [{ name = "Marcelo Cantos" }]
dependencies = [
    "mss>=9.0.1",
    "Pillow>=10.0.0",
    "ImageHash>=4.3.1",
]

[project.scripts]
pageflip = "pageflip.cli:main"

[project.urls]
Homepage = "https://github.com/marcelocantos/pageflip"
Repository = "https://github.com/marcelocantos/pageflip"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/pageflip"]
